@@ -1,27 +1,22 @@
-using System;
 using System.Text;
-using System.Threading.Tasks;
-using Haproxy.AgentCheck.Metrics;
+using Lucca.Infra.Haproxy.AgentCheck.Metrics;
 using Microsoft.AspNetCore.Connections;
 
-namespace Haproxy.AgentCheck.Endpoints
+namespace Lucca.Infra.Haproxy.AgentCheck.Endpoints;
+
+internal class TcpMiddleware(State state) : ConnectionHandler
 {
-    public class TcpMiddleware : ConnectionHandler
+    public override async Task OnConnectedAsync(ConnectionContext connection)
     {
-        private readonly StateProjection _stateProjection;
+        ArgumentNullException.ThrowIfNull(connection);
 
-        public TcpMiddleware(StateProjection computeLimits)
+        var up = state.IsUp ? "up" : "down";
+        if (!state.IsUp)
         {
-            _stateProjection = computeLimits;
+            up += $"#{state.BrokenCircuitsBreakers.Count} failures ({string.Join(",", state.BrokenCircuitsBreakers)})";
         }
 
-        public override async Task OnConnectedAsync(ConnectionContext connection)
-        {
-            if (connection == null)
-                throw new ArgumentNullException(nameof(connection));
-
-            await connection.Transport.Output.WriteAsync(Encoding.ASCII.GetBytes($"up {_stateProjection.Weight}%\n").AsMemory());
-            await connection.Transport.Output.FlushAsync();
-        }
+        await connection.Transport.Output.WriteAsync(Encoding.ASCII.GetBytes($"{state.Weight}%\n {up}").AsMemory(), connection.ConnectionClosed);
+        await connection.Transport.Output.FlushAsync(connection.ConnectionClosed);
     }
 }
