@@ -1,26 +1,51 @@
-using System;
-using Haproxy.AgentCheck.Hosting;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+#pragma warning disable CA1506
+using Lucca.Infra.Haproxy.AgentCheck.Config;
+using Lucca.Infra.Haproxy.AgentCheck.Endpoints;
+using Lucca.Infra.Haproxy.AgentCheck.Hosting;
+using Lucca.Infra.Haproxy.AgentCheck.Metrics;
+using Serilog;
+using Serilog.Core;
 
-namespace Haproxy.AgentCheck
+Environment.SetEnvironmentVariable("DD_TRACE_ENABLED", "0", EnvironmentVariableTarget.Process);
+var builder = WebApplication.CreateBuilder(args);
+builder.Logging.AddSerilog(CreateLoggerConfiguration(builder.Configuration));
+builder.Host.UseSystemService();
+builder.WebHost.UseKestrelOnPorts(8042, 4243);
+builder.Services.AddHostedService<ProcessCountersBackgroundService>();
+builder.Services.AddHostedService<SystemCountersBackgroundService>();
+builder.Services.AddOptions<WatchConfig>()
+    .BindConfiguration("Watch")
+    .ValidateOnStart()
+    .ValidateDataAnnotations();
+builder.Services.AddOptions<RulesConfig>()
+    .BindConfiguration("Rules")
+    .ValidateOnStart()
+    .ValidateDataAnnotations();
+builder.Services.AddMetricCollector();
+builder.Services.AddSingleton<State>();
+builder.Services.AddSingleton(TimeProvider.System);
+var app = builder.Build();
+app.MapGet("", HttpMiddleware.Invoke);
+await app.RunAsync();
+
+static Logger CreateLoggerConfiguration(IConfiguration configuration)
 {
-    public static class Program
-    {
-        public static void Main(string[] args)
-        {
-            Environment.SetEnvironmentVariable("DD_TRACE_ENABLED", "0", EnvironmentVariableTarget.Process);
-            CreateHostBuilder(args).Build().Run();
-        }
+    return new LoggerConfiguration()
+        .Enrich.WithProcessId()
+        .Enrich.WithProcessName()
+        .Enrich.WithAssemblyName()
+        .Enrich.WithAssemblyInformationalVersion()
+        .Enrich.WithMachineName()
+        .ReadFrom.Configuration(configuration)
+        .CreateLogger();
+}
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .UseSystemService()
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder
-                        .UseStartup<Startup>()
-                        .UseKestrelOnPorts(8042, 4243);
-                });
+#pragma warning restore CA1506
+#pragma warning disable S1118
+namespace Lucca.Infra.Haproxy.AgentCheck
+{
+    public partial class Program
+    {
     }
 }
+#pragma warning restore S1118
